pyproject.toml
[project]
name = "ai-http-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "aiohttp>=3.10.11",
    "python-dotenv>=1.0.1",
]
